requirements.txt
jira==3.1.1
crowdin-api-client==1.1.1